requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.31
psycopg2-binary==2.9.9
alembic==1.13.2
pydantic==2.8.2
python-dotenv==1.0.1
